requirements.txt
streamlit==1.33.0
pandas==2.2.1
requests==2.31.0
plotly==5.20.0
python-dotenv==1.0.1
langchain==0.1.16
langchain-openai==0.1.3